+++ CHANGELOG.md
@@ -2,12 +2,22 @@
 
 <hr>
 
+## [1.3.15] 2023-07-17
+
+### Added
+
+- API Core - Wait for following order state:
+- API Core - Verifies order state:
+- API Core - wait for order state:
+- API Core - Operator post Lazada 3PL using data below:
+
+<hr>
+
 ## [1.3.14] 2023-07-10
 
 ### Added
 
 - API Core - Update priority level of an order:
-- API Core - Operator post Lazada 3PL using data below:
 
 ### Deleted